send_start_message関数の切り出しと適用
